+++ frame/pablo/src/mock.rs
@@ -101,7 +101,7 @@ pub type Balance = u128;
 pub type AssetId = u128;
 pub type Amount = i128;
 pub type PoolId = u128;
-pub type RewardPoolId = u16;
+pub type RewardPoolId = u128;
 pub type PositionId = u128;
 
 parameter_type_with_key! {
